Выносит компоненты контейнеров из main в отдельные файлы
