+++ src/components/films-container.js
@@ -1,0 +1,4 @@
+// Возвращает разметку общего блока для карточек фильмов
+export const getFilmsContainerTemplate = () => {
+  return `<section class="films"></section>`.trim();
+};
