+++ src/components/filter-container.js
@@ -1,0 +1,4 @@
+// Возвращает разметку блока для фильтра
+export const getFilterContainerTemplate = () => {
+  return `<nav class="main-navigation"></nav>`.trim();
+};
