+++ src/components/sort-container.js
@@ -1,0 +1,4 @@
+// Возвращает разметку блока для сортировки
+export const getSortingContainerTemplate = () => {
+  return `<ul class="sort"></ul>`.trim();
+};
