+++ src/components/stats-elem-template.js
@@ -1,0 +1,4 @@
+// Возвращает разметку пункта фильтра Stats
+export const getStatsElemTemplate = () => {
+  return `<a href="#stats" class="main-navigation__item main-navigation__item--additional">Stats</a>`.trim();
+};
